Decompose ado.Register God Function (528 LOC, complexity 26→5)
Closes #120.

Extract 20 tool specs from ado.Register into 5 per-category registerX
functions co-located with domain files, following ADR-028 recursively:

- registerPullRequests → pr.go (6 specs)
- registerPipelines   → pipeline.go (7 specs)
- registerBuilds      → build.go [NEW] (3 specs)
- registerRepository  → repo.go [NEW] (2 specs)
- registerPolicy      → policy.go (2 specs)

ado.Register is now a 59-line orchestrator (complexity 5) that constructs
the ADO manager and formatter once, then delegates to all 5 registerX
functions via a loop.

Promote marshalIndentResult and appendTruncationNote from inline closures
to package-level helpers in register.go.

All registerX functions are unexported, share the uniform signature
(tools.Registry, *AdoManager, PipelineFormatter) error, and are called
only from ado.Register.
