+++ internal/tools/integrations/ado/pipeline.go
@@ -4,6 +4,7 @@
 package ado
 
 import (
+	"context"
 	"fmt"
 	"strings"
 
@@ -310,4 +311,229 @@ type CreatePipelineResult struct {
 type UpdateVariablesResult struct {
 	Cancelled    bool
 	DefinitionID int
+}
+
+// registerPipelines registers the Pipeline category tools (list, run, definition, logs, create).
+func registerPipelines(r tools.Registry, m *AdoManager, f PipelineFormatter) error {
+	type toolSpec struct {
+		decl    *tools.ToolDeclaration
+		handler tools.ToolFunc
+	}
+
+	specs := []toolSpec{
+		{
+			decl: &tools.ToolDeclaration{
+				Name:        "ado_list_pipelines",
+				Description: "Lists pipeline names and their corresponding IDs in an Azure DevOps project.",
+				Parameters: &tools.Schema{
+					Type: "OBJECT",
+					Properties: map[string]*tools.Schema{
+						"organization": {Type: "STRING", Description: "The Azure DevOps organization name."},
+						"project":      {Type: "STRING", Description: "The project name or ID."},
+					},
+					Required: []string{"organization", "project"},
+				},
+			},
+			handler: func(ctx context.Context, args map[string]interface{}, hb chan<- struct{}) (tools.ToolResult, error) {
+				pipelines, err := m.ListPipelines(ctx, args)
+				if err != nil {
+					return tools.ToolResult{}, err
+				}
+				return tools.ToolResult{Text: f.FormatPipelineList(pipelines)}, nil
+			},
+		},
+		{
+			decl: &tools.ToolDeclaration{
+				Name:        "ado_list_pipeline_runs",
+				Description: "Lists recent runs (executions) for a specific Azure DevOps pipeline.",
+				Parameters: &tools.Schema{
+					Type: "OBJECT",
+					Properties: map[string]*tools.Schema{
+						"organization":  {Type: "STRING", Description: "The Azure DevOps organization name."},
+						"project":       {Type: "STRING", Description: "The project name or ID."},
+						"pipeline_id":   {Type: "INTEGER", Description: "The ID of the pipeline definition."},
+						"pipeline_name": {Type: "STRING", Description: "The name of the pipeline (used if pipeline_id is omitted)."},
+						"repository":    {Type: "STRING", Description: "Filter runs by repository ID (UUID) or exact internal reference name, as raw names are often omitted in run resource metadata."},
+						"top":           {Type: "INTEGER", Description: "Maximum number of runs to return (default 10)."},
+					},
+					Required: []string{"organization", "project"},
+				},
+			},
+			handler: func(ctx context.Context, args map[string]interface{}, hb chan<- struct{}) (tools.ToolResult, error) {
+				pipelineID, runs, err := m.ListPipelineRuns(ctx, args)
+				if err != nil {
+					return tools.ToolResult{}, err
+				}
+				return tools.ToolResult{Text: f.FormatPipelineRunsList(pipelineID, runs)}, nil
+			},
+		},
+		{
+			decl: &tools.ToolDeclaration{
+				Name:        "ado_get_pipeline_run",
+				Description: "Retrieves the detailed status and result of a specific Azure DevOps pipeline run.",
+				Parameters: &tools.Schema{
+					Type: "OBJECT",
+					Properties: map[string]*tools.Schema{
+						"organization": {Type: "STRING", Description: "The Azure DevOps organization name."},
+						"project":      {Type: "STRING", Description: "The project name or ID."},
+						"pipeline_id":  {Type: "INTEGER", Description: "The ID of the pipeline definition."},
+						"run_id":       {Type: "INTEGER", Description: "The ID of the specific execution."},
+					},
+					Required: []string{"organization", "project", "pipeline_id", "run_id"},
+				},
+			},
+			handler: func(ctx context.Context, args map[string]interface{}, hb chan<- struct{}) (tools.ToolResult, error) {
+				run, err := m.GetPipelineRun(ctx, args)
+				if err != nil {
+					return tools.ToolResult{}, err
+				}
+				return tools.ToolResult{Text: f.FormatPipelineRunDetail(run)}, nil
+			},
+		},
+		{
+			decl: &tools.ToolDeclaration{
+				Name:        "ado_get_pipeline_definition",
+				Description: "Retrieves the full configuration metadata of an existing Azure DevOps pipeline, including variables and security settings.",
+				Parameters: &tools.Schema{
+					Type: "OBJECT",
+					Properties: map[string]*tools.Schema{
+						"organization": {Type: "STRING", Description: "The Azure DevOps organization name."},
+						"project":      {Type: "STRING", Description: "The project name or ID."},
+						"pipeline_id":  {Type: "INTEGER", Description: "The numeric ID of the pipeline to inspect."},
+					},
+					Required: []string{"organization", "project", "pipeline_id"},
+				},
+			},
+			handler: func(ctx context.Context, args map[string]interface{}, hb chan<- struct{}) (tools.ToolResult, error) {
+				def, err := m.GetPipelineDefinition(ctx, args)
+				if err != nil {
+					return tools.ToolResult{}, err
+				}
+				return marshalIndentResult(def)
+			},
+		},
+		{
+			decl: &tools.ToolDeclaration{
+				Name:        "ado_get_pipeline_logs",
+				Description: "Fetches the console output/logs of a failed pipeline run for diagnosis.",
+				Parameters: &tools.Schema{
+					Type: "OBJECT",
+					Properties: map[string]*tools.Schema{
+						"organization":  {Type: "STRING", Description: "The Azure DevOps organization name."},
+						"project":       {Type: "STRING", Description: "The project name or ID."},
+						"pipeline_id":   {Type: "INTEGER", Description: "The ID of the pipeline definition."},
+						"run_id":        {Type: "INTEGER", Description: "The ID of the specific execution."},
+						"log_id":        {Type: "INTEGER", Description: "The specific log ID. If omitted, the tool lists available logs."},
+						"tail_lines":    {Type: "INTEGER", Description: "Return the last N lines. Default: 200"},
+						"head_lines":    {Type: "INTEGER", Description: "Return the first N lines (optional)."},
+						"filter_query":  {Type: "STRING", Description: "Regex to filter log lines (e.g. 'error|failed'). WARNING: This parameter OVERRIDES tail/head/pagination. It scans the ENTIRE file. Do NOT use greedy wildcards like '.*' as it will return massive payloads and crash the context."},
+						"context_lines": {Type: "INTEGER", Description: "Lines of context around filter matches. Default: 5"},
+						"start_line":    {Type: "INTEGER", Description: "Start line for pagination."},
+						"max_lines":     {Type: "INTEGER", Description: "Maximum lines for pagination."},
+					},
+					Required: []string{"organization", "project", "pipeline_id", "run_id"},
+				},
+			},
+			handler: func(ctx context.Context, args map[string]interface{}, hb chan<- struct{}) (tools.ToolResult, error) {
+				var dispatch struct {
+					LogId int `json:"log_id"`
+				}
+				_ = tools.UnmarshalArgs(args, &dispatch)
+
+				if dispatch.LogId == 0 {
+					runID, logs, err := m.ListPipelineLogs(ctx, args)
+					if err != nil {
+						return tools.ToolResult{}, err
+					}
+					return tools.ToolResult{Text: f.FormatPipelineLogList(runID, logs)}, nil
+				}
+
+				content, err := m.GetPipelineLogContent(ctx, args, hb)
+				if err != nil {
+					return tools.ToolResult{}, err
+				}
+				return tools.ToolResult{Text: appendTruncationNote(content)}, nil
+			},
+		},
+		{
+			decl: &tools.ToolDeclaration{
+				Name:        "ado_create_pipeline",
+				Description: "Creates a new Azure DevOps YAML build pipeline. Idempotent: returns existing ID if the name already exists. Triggers security confirmation.",
+				Parameters: &tools.Schema{
+					Type: "OBJECT",
+					Properties: map[string]*tools.Schema{
+						"organization":    {Type: "STRING", Description: "The Azure DevOps organization name."},
+						"project":         {Type: "STRING", Description: "The project name or ID."},
+						"name":            {Type: "STRING", Description: "The desired name of the new pipeline."},
+						"repository_id":   {Type: "STRING", Description: "The UUID of the Git repository containing the YAML."},
+						"yaml_path":       {Type: "STRING", Description: "The exact path to the YAML file in the repository (e.g., '/dev-3/caddy/cicd/main.yaml')."},
+						"variable_groups": {Type: "ARRAY", Description: "A list of numeric IDs for existing Variable Groups (e.g., 'WebSC-Common-Vars') to be linked to the new pipeline.", Items: &tools.Schema{Type: "INTEGER"}},
+						"variables":       {Type: "OBJECT", Description: "A collection of key-value pairs (name -> {value, isSecret, allowOverride}) to be stored directly in the pipeline's 'Variables' configuration. 'allowOverride' defaults to true, letting users change the value during a run."},
+					},
+					Required: []string{"organization", "project", "name", "repository_id", "yaml_path"},
+				},
+			},
+			handler: func(ctx context.Context, args map[string]interface{}, hb chan<- struct{}) (tools.ToolResult, error) {
+				result, err := m.CreatePipeline(ctx, args)
+				if err != nil {
+					return tools.ToolResult{}, err
+				}
+				switch {
+				case result.AlreadyExisted:
+					return tools.ToolResult{Text: fmt.Sprintf("Pipeline '%s' already exists with ID: %d", result.Name, result.PipelineID)}, nil
+				case result.Cancelled:
+					return tools.ToolResult{Text: "Pipeline creation cancelled by user."}, nil
+				default:
+					return tools.ToolResult{Text: fmt.Sprintf("Successfully created pipeline '%s' with ID: %d", result.Name, result.PipelineID)}, nil
+				}
+			},
+		},
+		{
+			decl: &tools.ToolDeclaration{
+				Name:        "ado_run_pipeline",
+				Description: "Triggers a manual run of an existing Azure DevOps YAML pipeline. Returns the Run ID for tracking. Triggers security confirmation.",
+				Parameters: &tools.Schema{
+					Type: "OBJECT",
+					Properties: map[string]*tools.Schema{
+						"organization":        {Type: "STRING", Description: "The Azure DevOps organization name."},
+						"project":             {Type: "STRING", Description: "The project name or ID."},
+						"pipeline_id":         {Type: "INTEGER", Description: "The numeric ID of the pipeline to run."},
+						"branch":              {Type: "STRING", Description: "The branch or tag to build from (e.g., 'main', 'v1.2.0'). Full refs like 'refs/tags/v1.0' are also supported."},
+						"template_parameters": {Type: "OBJECT", Description: "Key-value string pairs for YAML parameters."},
+						"variables":           {Type: "OBJECT", Description: "Key-value string pairs for pipeline variables."},
+					},
+					Required: []string{"organization", "project", "pipeline_id"},
+				},
+			},
+			handler: func(ctx context.Context, args map[string]interface{}, hb chan<- struct{}) (tools.ToolResult, error) {
+				branchRaw, _ := args["branch"].(string)
+				refName := f.FormatBranchRef(branchRaw)
+
+				argsCopy := make(map[string]interface{}, len(args)+1)
+				for k, v := range args {
+					argsCopy[k] = v
+				}
+				argsCopy["_ref_name"] = refName
+
+				result, err := m.RunPipeline(ctx, argsCopy)
+				if err != nil {
+					return tools.ToolResult{}, err
+				}
+				if result.Cancelled {
+					return tools.ToolResult{Text: "Pipeline run cancelled by user."}, nil
+				}
+				return tools.ToolResult{
+					Text: fmt.Sprintf("Successfully triggered pipeline run ID: %d\nWeb URL: %s", result.RunID, result.WebURL),
+				}, nil
+			},
+		},
+	}
+
+	for _, spec := range specs {
+		if err := r.RegisterToToolkit("ado", spec.decl, spec.handler); err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
